Api Provada

+++ ShopApp.Domain/Models/OrderDetails/OrderDetailsBaseModel/OrderDetailsModel.cs
@@ -7,7 +7,7 @@ namespace ShopApp.Domain.Models.OrderDetails.OrderDetailsBaseModel
         public int orderid { get; set; }
         public int productid { get; set; }
         public decimal unitprice { get; set; }
-        public int qty { get; set; }
+        public decimal qty { get; set; }
         public decimal discount { get; set; }
     }
 }
